refactor: Servlet 구조 단순화 개선 및 스프링 맥락 제거 개선
