+++ tomcat/src/main/java/com/techcourse/servlet/GreetingServlet.java
@@ -6,7 +6,6 @@ import org.apache.coyote.http11.response.HttpServletResponse;
 
 public class GreetingServlet extends HttpServlet {
 
-
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) {
         response.ok("Hello world!", "plain");
