--- tomcat/src/main/java/com/techcourse/servlet/mapping/HandlerMapping.java
@@ -1,11 +1,0 @@
-package com.techcourse.servlet.mapping;
-
-import org.apache.catalina.servlet.Servlet;
-import org.apache.coyote.http11.request.HttpServletRequest;
-
-public interface HandlerMapping {
-
-    boolean hasHandlerFor(HttpServletRequest httpServletRequest);
-
-    Servlet getHandler(HttpServletRequest httpServletRequest);
-}
