--- tomcat/src/test/java/com/techcourse/servlet/DispatcherServletTest.java
@@ -1,17 +1,0 @@
-package com.techcourse.servlet;
-
-import org.junit.jupiter.api.DisplayName;
-
-@DisplayName("디스패쳐 서블릿 테스트")
-class DispatcherServletTest {
-
-    private final DispatcherServlet dispatcherServlet = DispatcherServlet.getInstance();
-
-//    @DisplayName("핸들러 매핑 순서는 정적 자원 핸들러 -> 요청 정보 매핑 핸들러 순이다")
-//    @Test
-//
-//    @DisplayName("요청에 맞추어 요청을 처리할 수 있는 핸들러를 조회할 수 있다")
-//
-//    @DisplayName("")
-
-}
