+++ tomcat/src/test/java/com/techcourse/servlet/mapping/RequestMappingTest.java
@@ -1,0 +1,50 @@
+package com.techcourse.servlet.mapping;
+
+import static org.apache.coyote.http11.common.HttpProtocol.HTTP_11;
+import static org.apache.coyote.http11.request.line.Method.GET;
+import static org.assertj.core.api.Assertions.assertThat;
+
+import com.techcourse.servlet.DefaultServlet;
+import com.techcourse.servlet.LoginServlet;
+import org.apache.coyote.http11.common.HttpHeaders;
+import org.apache.coyote.http11.common.HttpMessageBody;
+import org.apache.coyote.http11.common.HttpProtocol;
+import org.apache.coyote.http11.request.HttpServletRequest;
+import org.apache.coyote.http11.request.line.Method;
+import org.apache.coyote.http11.request.line.RequestLine;
+import org.apache.coyote.http11.request.line.Uri;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+@DisplayName("요청 매핑 핸들러 조회 테스트")
+class RequestMappingTest {
+
+    private final RequestMapping requestMapping = new RequestMapping();
+
+    @DisplayName("매핑 정보가 일치하는 핸들러를 가져올 수 있다")
+    @Test
+    void getHandler() {
+        // given
+        RequestLine requestLine = new RequestLine(GET, new Uri("/login"), HTTP_11);
+        HttpServletRequest httpServletRequest = new HttpServletRequest(
+                requestLine,
+                new HttpHeaders(),
+                HttpMessageBody.createEmptyBody()
+        );
+        // when & then
+        assertThat(requestMapping.getServlet(httpServletRequest)).isInstanceOf(LoginServlet.class);
+    }
+
+    @DisplayName("매핑 정보가 일치하는 핸들러가 없다면 디폴트 서블릿이 반환된다")
+    @Test
+    void getDefaultHandler() {
+        // given
+        RequestLine requestLine = new RequestLine(Method.GET, new Uri("/index.html"), HttpProtocol.HTTP_11);
+        HttpHeaders httpHeaders = new HttpHeaders();
+        HttpMessageBody httpMessageBody = HttpMessageBody.createEmptyBody();
+        HttpServletRequest httpServletRequest = new HttpServletRequest(requestLine, httpHeaders, httpMessageBody);
+
+        // when & then
+        assertThat(requestMapping.getServlet(httpServletRequest)).isInstanceOf(DefaultServlet.class);
+    }
+}
